requirements.txt
playwright==1.38.0
pytest-playwright==0.4.2
pyyaml==6.0.1
python-dotenv==1.0.0
apscheduler==3.10.1
aiohttp==3.8.5
pillow==10.0.1
requests==2.31.0
pytesseract==0.3.10